backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.9.0",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "chromadb>=1.0.15",
    "click>=8.2.1",
    "fastapi>=0.115.14",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "langchain-text-splitters>=0.3.8",
    "langfuse>=3.1.3",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyflakes>=3.4.0",
    "pypdf2>=3.0.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "starlette>=0.46.2",
    "uvicorn>=0.35.0",
]
